[优化代码] 优化应用代码

+++ platform-sms-admin/admin-web/src/main/java/com/courage/platform/sms/admin/api/admin/AppInfoController.java
@@ -3,15 +3,19 @@ package com.courage.platform.sms.admin.api.admin;
 import com.courage.platform.sms.admin.common.utils.ResponseEntity;
 import com.courage.platform.sms.admin.domain.TSmsAppinfo;
 import com.courage.platform.sms.admin.domain.vo.Pager;
+import com.courage.platform.sms.admin.domain.vo.RecordVO;
 import com.courage.platform.sms.admin.service.AppInfoService;
+import com.courage.platform.sms.admin.service.SmsRecordService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -24,6 +28,9 @@ public class AppInfoController {
 
     @Autowired
     private AppInfoService appInfoService;
+
+    @Autowired
+    private SmsRecordService smsRecordService;
 
     @PostMapping(value = "/appList")
     public ResponseEntity<Pager> appList(String appkey, int page, int size) {
@@ -54,6 +61,12 @@ public class AppInfoController {
         if (Long.valueOf(id) == 1) {
             logger.info("默认应用不能删除");
             return ResponseEntity.fail("默认应用不能删除");
+        }
+        //判断应用是否有短信
+        List<RecordVO> recordVOList = smsRecordService.queryOneRecordVOByAppId(id);
+        if (!CollectionUtils.isEmpty(recordVOList)) {
+            logger.info("本应用曾发送过短信，不能删除");
+            return ResponseEntity.fail("本应用曾发送过短信，不能删除");
         }
         return appInfoService.deleteAppInfo(id);
     }
